Play around with JS modularization.
