+++ assets/js/utils/chosen.js
@@ -1,0 +1,11 @@
+/* global edd_vars */
+
+export const chosenVars = {
+	disable_search_threshold:  13,
+	search_contains:           true,
+	inherit_select_classes:    true,
+	single_backstroke_delete:  false,
+	placeholder_text_single:   edd_vars.one_option,
+	placeholder_text_multiple: edd_vars.one_or_more_option,
+	no_results_text:           edd_vars.no_results_text
+};
