+++ babel.config.js
@@ -1,0 +1,7 @@
+module.exports = function( api ) {
+	api.cache( true );
+
+	return {
+		presets: [ '@wordpress/default' ],
+	};
+};
