+++ package.json
@@ -1,22 +1,49 @@
 {
-  "name": "easy-digital-downloads",
-  "version": "2.9.6",
-  "private": true,
-  "devDependencies": {
-    "grunt": "^1.0.3",
-    "grunt-checktextdomain": "^1.0.1",
-    "grunt-contrib-clean": "^1.1.0",
-    "grunt-contrib-compress": "^1.4.3",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-cssmin": "^2.2.1",
-    "grunt-contrib-uglify": "^3.3.0",
-    "grunt-force-task": "^2.0.0",
-    "grunt-rtlcss": "~2.0.1",
-    "grunt-wp-i18n": "^1.0.2",
-    "load-grunt-tasks": "^3.5.2",
-    "makepot": "^0.3.0"
-  },
-  "dependencies": {
-    "npm": "^6.3.0"
-  }
+	"name": "easy-digital-downloads",
+	"version": "2.9.6",
+	"description": "The easiest way to sell digital products with WordPress.",
+	"private": true,
+	"license": "GPL-2.0-or-later",
+	"main": "webpack.config.js",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/easydigitaldownloads/easy-digital-downloads/"
+	},
+	"bugs": {
+		"url": "https://github.com/easydigitaldownloads/easy-digital-downloads/issues"
+	},
+	"homepage": "https://easydigitaldownloads.com",
+	"engines": {
+		"node": ">=8.0.0",
+		"npm": ">=6.0.0"
+	},
+	"devDependencies": {
+		"@babel/core": "7.0.0",
+		"@wordpress/babel-preset-default": "2.0.2",
+		"babel-eslint": "8.0.3",
+		"babel-loader": "8.0.0",
+		"cross-env": "3.2.4",
+		"eslint": "4.16.0",
+		"eslint-config-wordpress": "2.0.0",
+		"grunt": "^1.0.3",
+		"grunt-checktextdomain": "^1.0.1",
+		"grunt-contrib-clean": "^1.1.0",
+		"grunt-contrib-compress": "^1.4.3",
+		"grunt-contrib-copy": "^1.0.0",
+		"grunt-contrib-cssmin": "^2.2.1",
+		"grunt-contrib-uglify": "^3.3.0",
+		"grunt-force-task": "^2.0.0",
+		"grunt-rtlcss": "~2.0.1",
+		"grunt-wp-i18n": "^1.0.2",
+		"load-grunt-tasks": "^3.5.2",
+		"makepot": "^0.3.0",
+		"webpack": "4.8.3",
+		"webpack-cli": "2.1.3"
+	},
+	"scripts": {
+		"build": "cross-env NODE_ENV=production webpack",
+		"dev": "cross-env NODE_ENV=default webpack --watch",
+		"lint": "./node_modules/.bin/eslint .; exit 0",
+		"lint:fix": "./node_modules/.bin/eslint . --fix; exit 0"
+	}
 }
